% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{multi,character,integer-method}
\alias{multi,character,integer-method}
\title{S7 multi-dispatch method}
\usage{
## S7 method for classes <character>, <integer>
multi(x, y, ...)
}
\description{
S7 multi-dispatch method
}
