% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{multi}
\alias{multi}
\alias{multi,integer,character-method}
\title{An S7 multi-dispatch generic}
\usage{
multi(x, y, ...)

## S7 method for classes <integer>, <character>
multi(x, y, ...)
}
\arguments{
\item{x, y}{A parameter}
}
\description{
An S7 multi-dispatch generic
}
