% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{range}
\alias{range}
\title{Range Class}
\usage{
range(start = class_missing, end = class_missing)
}
\description{
A class representing a range with start and end values.
}
\section{Slots}{

\describe{
\item{\code{start}}{A numeric value representing the start of the range.}

\item{\code{end}}{A numeric value representing the end of the range.}
}}

\section{Validation}{

The class has a validator function that checks if the start and end values are of length 1 and if the end value is greater than or equal to the start value.
}

\examples{

range <- new("range", start = 1, end = 10)
range

}
